Move useReadQuery to own hooks file
