+++ src/react/hooks/__tests__/useBackgroundQuery.test.tsx
@@ -35,7 +35,8 @@ import {
   mockSingleLink,
 } from '../../../testing';
 import { concatPagination, offsetLimitPagination } from '../../../utilities';
-import { useBackgroundQuery, useReadQuery } from '../useBackgroundQuery';
+import { useBackgroundQuery } from '../useBackgroundQuery';
+import { useReadQuery } from '../useReadQuery';
 import { ApolloProvider } from '../../context';
 import { QUERY_REFERENCE_SYMBOL } from '../../cache/QueryReference';
 import { SuspenseCache } from '../../cache';
