+++ src/react/hooks/useReadQuery.ts
@@ -1,0 +1,55 @@
+import { useState, useMemo, useEffect } from "react";
+import invariant from "ts-invariant";
+import { NetworkStatus } from "../../core";
+import { QUERY_REFERENCE_SYMBOL, type QueryReference } from "../cache/QueryReference";
+import { __use } from "./internal";
+import { toApolloError } from "./useSuspenseQuery";
+
+export function useReadQuery<TData>(queryRef: QueryReference<TData>) {
+  const [, forceUpdate] = useState(0);
+  const internalQueryRef = queryRef[QUERY_REFERENCE_SYMBOL];
+  invariant(
+    internalQueryRef.promiseCache,
+    'It appears that `useReadQuery` was used outside of `useBackgroundQuery`. ' +
+      '`useReadQuery` is only supported for use with `useBackgroundQuery`. ' +
+      'Please ensure you are passing the `queryRef` returned from `useBackgroundQuery`.'
+  );
+
+  const skipResult = useMemo(() => {
+    const error = toApolloError(internalQueryRef.result);
+
+    return {
+      loading: false,
+      data: internalQueryRef.result.data,
+      networkStatus: error ? NetworkStatus.error : NetworkStatus.ready,
+      error,
+    };
+  }, [internalQueryRef.result]);
+
+  let promise = internalQueryRef.promiseCache.get(internalQueryRef.key);
+
+  if (!promise) {
+    promise = internalQueryRef.promise;
+    internalQueryRef.promiseCache.set(internalQueryRef.key, promise);
+  }
+
+  useEffect(() => {
+    return internalQueryRef.listen((promise) => {
+      internalQueryRef.promiseCache!.set(internalQueryRef.key, promise);
+      forceUpdate((prevState) => prevState + 1);
+    });
+  }, [queryRef]);
+
+  const result =
+    internalQueryRef.watchQueryOptions.fetchPolicy === 'standby'
+      ? skipResult
+      : __use(promise);
+
+  return useMemo(() => {
+    return {
+      data: result.data,
+      networkStatus: result.networkStatus,
+      error: toApolloError(result),
+    };
+  }, [result]);
+}
